change admin console to support logs print
