+++ cmd/auto-complete.go
@@ -409,14 +409,17 @@ var completeCmds = map[string]complete.Predictor{
 	"/alias/remove": aliasCompleter,
 	"/alias/import": nil,
 
-	"/support/callhome": aliasCompleter,
-	"/support/logs":     aliasCompleter,
-	"/support/register": aliasCompleter,
-	"/support/diag":     aliasCompleter,
-	"/support/profile":  aliasCompleter,
-	"/support/inspect":  aliasCompleter,
-	"/support/perf":     aliasCompleter,
-	"/support/metrics":  aliasCompleter,
+	"/support/callhome":     aliasCompleter,
+	"/support/logs/enable":  aliasCompleter,
+	"/support/logs/disable": aliasCompleter,
+	"/support/logs/status":  aliasCompleter,
+	"/support/logs/print":   aliasCompleter,
+	"/support/register":     aliasCompleter,
+	"/support/diag":         aliasCompleter,
+	"/support/profile":      aliasCompleter,
+	"/support/inspect":      aliasCompleter,
+	"/support/perf":         aliasCompleter,
+	"/support/metrics":      aliasCompleter,
 
 	"/update": nil,
 }
